refactor 패키지 정리
* 레거시 제거
* 패키지 구분
