--- src/main/java/core/mvc/asis/RequestMapping.java
@@ -1,30 +1,0 @@
-package core.mvc.asis;
-
-import core.mvc.HandlerMapping;
-import javax.servlet.http.HttpServletRequest;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public class RequestMapping implements HandlerMapping {
-    private static final Logger logger = LoggerFactory.getLogger(DispatcherServlet.class);
-    private Map<String, Controller> mappings = new HashMap<>();
-
-    void initMapping() {
-        logger.info("Initialized Request Mapping!");
-        mappings.keySet().forEach(path -> {
-            logger.info("Path : {}, Controller : {}", path, mappings.get(path).getClass());
-        });
-    }
-
-    void put(String url, Controller controller) {
-        mappings.put(url, controller);
-    }
-
-    @Override
-    public Controller getHandler(HttpServletRequest request) {
-        return mappings.get(request.getRequestURI());
-    }
-}
