+++ src/main/java/core/mvc/resolver/BeanHandlerMethodArgumentResolver.java
@@ -4,11 +4,10 @@ import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Parameter;
-import java.lang.reflect.Type;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class BeanHandlerMethodArgumentResolver implements HandlerMethodArgumentResolver {
+class BeanHandlerMethodArgumentResolver implements HandlerMethodArgumentResolver {
 
     @Override
     public boolean supportsParameter(MethodParameter method) {
