+++ src/test/java/core/mvc/HandlerMethodArgumentResolverTest.java
@@ -1,6 +1,5 @@
-package core.mvc.tobe;
+package core.mvc;
 
-import core.mvc.ModelAndView;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
