+++ src/test/java/core/mvc/MyController.java
@@ -1,10 +1,9 @@
-package core.mvc.tobe;
+package core.mvc;
 
 import core.annotation.web.Controller;
 import core.annotation.web.RequestMapping;
 import core.annotation.web.RequestMethod;
 import core.db.DataBase;
-import core.mvc.ModelAndView;
 import next.model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
