+++ src/test/java/core/mvc/TestUserController.java
@@ -1,9 +1,8 @@
-package core.mvc.tobe;
+package core.mvc;
 
 import core.annotation.web.PathVariable;
 import core.annotation.web.RequestMapping;
 import core.annotation.web.RequestMethod;
-import core.mvc.ModelAndView;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
