+++ src/test/java/core/mvc/resolver/PathHandlerMethodArgumentResolverTest.java
@@ -2,7 +2,7 @@ package core.mvc.resolver;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import core.mvc.tobe.TestUserController;
+import core.mvc.TestUserController;
 import java.lang.reflect.Method;
 import java.util.Arrays;
 import org.junit.jupiter.api.Test;
